Move render_create_index to SqlRenderer
